tests: Move filter processing into a separate file
The file is currently used only by FilterJSCompiler tests but this will
change soon.
